+++ src/models/attributes/index.ts
@@ -78,9 +78,8 @@ export class Attributes {
       value,
     };
 
-    const query = `SELECT attributes.*
-                   FROM attribute_keys
-                   JOIN attributes ON attribute_keys.id = attributes.attribute_key_id
+    const query = `SELECT *
+                   FROM attributes
                    WHERE collection_id = $/collectionId/
                    AND key = $/key/
                    AND value = $/value/`;